Bug correction and deletion of useless test class #30 #11
